Extend CDASegmentAnnotator to tag the section headings.
Previously, CDASegmentAnnotator was just tagging the section bodies (i.e.
the regions between the identified section headings).  This change extends
it to tag the headings as well.

This adds a new class to the typesystem:
org.apache.ctakes.typesystem.type.textspan.SectionHeading.

This includes a change to the code that tags the sections by looking at the
gaps between section headings.  The code wasn't doing what it was claiming
because the code commented as "handle case for first section" wasn't
actually assigning its computed value to anything.  This meant that it
would always start the section Segment at the final character of the text
identified as a heading.  This meant that Segments would only cover the
section body text rather than the heading plus the body as it claimed to be
doing.  This code changes so that we tag a SectionHeading, followed by a
Segment for the section body, followed by the next SectionHeading and so on.
These are disjoint.

The annotation positions are moved inwards to avoid tagging any whitespace
at either end.

This also moves the handling for the SIMPLE_SEGMENT case out of the final
loop, because it was pointlessly complicated handling it there.

The TODO comment above the segment sorting goes away.  It wasn't a TODO,
despite claiming so, because it was an inherent part of the algorithm.

This also includes the fix from CTAKES-450, plus the corresponding tests.
That work is by Kean Kaufmann and Alex Zbarcea.