Fix if the database is in a URL-unsafe path.
Fix JdbcConnectionFactory in the case where it uses FileLocator to find
the database, but the path would be encoded inside a URL (e.g. if it has
a space in it).  In this case, getConnectionUrl was producing a
percent-encoded URL, but Hsqldb doesn't decode the URL it receives, it just
treats it like a scheme followed by a plain file path.

Fix this by returning the plain string form of the file path from
getConnectionUrl, and rename it to resolveDatabasePath because it no longer
returns a URL.