CTAKES-505: JdbcConnectionFactory does not handle URLs with properties
JdbcConnectionFactory includes code to resolve the reference to the
hsqldb file by appending the .script extension and searching for the
file using FileLocator. This fails if the JDBC URL has properties on
the end (e.g. .../sno_rx_16ab/sno_rx_16ab;readonly=true) because
JdbcConnectionFactory doesn't handle the properties and so it searches
for an invalid file.

Fix this by splitting the properties from the URL before calling
FileLocator, and then putting them back on the resulting path.