Extend CDASegmentAnnotator to ignore text that trails a section.
CDASegmentAnnotator gains a new parameter, PARAM_SECTION_END_MARKERS.
This can be set by the caller to a list of strings giving possible
markers for the end of a section.  This might be things like "Signed by"
or whatever suits the usecase.  Any section that contains that text will
have the Segment extent brought in so that the "Signed by ..." text is
not covered.