Change the regexes in CDASegmentAnnotator.
No longer look for "<CDA code>" as a header -- I'm not seeing this in any
use case, and it runs the risk of false positives with other numeric codes.
This means that we skip immediately to column 2 of ccda_sections.txt.

No longer accept "<Header>-" as a header -- I'm not seeing this in any use
case.

No longer accept "<Header>  " as a header; instead require a newline to
terminate (with optional trailing space and colon).

Allow \u2003 (em-space) as a whitespace character.