Add a "Term processing mode" option to the DictionaryCreator UI.
This means that users can specify a different set of term processing logic
other than the "tiny" set that was used before (and is now the default in
the UI).

As part of this, don't exit if reading one of the files fails -- just throw
the IOException instead.  This means that we simply behave as if the file
was empty rather than bailing.  The "default" set has no UnwantedTexts.txt,
for example.