CTAKES-520: SentenceDetectorAnnotatorBIO token scanning performance issues
Fix some performance issues in SentenceDetectorAnnotatorBIO.

SentenceDetectorAnnotatorBIO iterates over every character in the Segment
and classifies it as Begin, Inside, or Outside a Sentence. When doing this,
it needs to know the next and previous token from the current character.

Before this change, it found these tokens afresh for each character. That
meant that it started from the current character, and scanned forward and
backwards looking for whitespace until it finds the boundaries of the tokens
either side of the current position. This is very wasteful; when the current
index steps within a word, the tokens do not change since we're still within
the same word. Also, since we're scanning in one direction, we never need to
scan for the previous token, because we already know it.

This fixes the problem by keeping track of the word boundary, and only
scanning for the next token when we have reached the boundary of the current
one. Also, the previous token is simply taken as the token from the previous
iteration.

The token features don't change unless the current token changes, so we
cache those too.